Add NavigationPixelTabExtension
