+++ macOS/DuckDuckGo/Tab/Model/Tab+Navigation.swift
@@ -45,6 +45,7 @@ extension Tab: NavigationResponder {
             .weak(nullable: self.aiChat),
 
             .weak(nullable: self.navigationHotkeyHandler),
+            .weak(nullable: self.navigationPixel),
             .weak(nullable: self.brokenSiteInfo),
             .weak(nullable: self.tabCrashRecovery),
