+++ macOS/DuckDuckGo/Tab/Model/Tab.swift
@@ -1326,24 +1326,6 @@ extension Tab/*: NavigationResponder*/ { // to be moved to Tab+Navigation.swift
         permissions.tabDidStartNavigation()
         userInteractionDialog = nil
 
-        let shouldFireNavigationPixel: Bool = {
-            /// Fire navigation pixel on all navigations except for loading error pages
-            if navigation.navigationAction.navigationType == .alternateHtmlLoad {
-                return false
-            }
-            /// Sometimes navigation type for an error page is reported as `.other`, so checking also target frame URL
-            /// This has a side effect of filtering out also some navigations starting on an error page (e.g. using a reload button,
-            /// that is also reported as `.other`).
-            if navigation.navigationAction.navigationType == .other && navigation.navigationAction.targetFrame?.url == .error {
-                return false
-            }
-            return true
-        }()
-
-        if shouldFireNavigationPixel {
-            PixelKit.fire(GeneralPixel.navigation)
-        }
-
         // Unnecessary assignment triggers publishing
         if error != nil,
            navigation.navigationAction.navigationType != .alternateHtmlLoad { // error page navigation
@@ -1367,9 +1349,6 @@ extension Tab/*: NavigationResponder*/ { // to be moved to Tab+Navigation.swift
     func navigation(_ navigation: Navigation, didSameDocumentNavigationOf navigationType: WKSameDocumentNavigationType) {
         guard navigation.isCurrent else { return }
 
-        if navigationType != .sessionStateReplace {
-            PixelKit.fire(GeneralPixel.navigation)
-        }
         invalidateInteractionStateData()
     }
