+++ macOS/DuckDuckGo/Tab/TabExtensions/TabExtensions.swift
@@ -218,6 +218,9 @@ extension TabExtensionsBuilder {
         add {
             NavigationHotkeyHandler(isTabPinned: args.isTabPinned, isBurner: args.isTabBurner)
         }
+        add {
+            NavigationPixelTabExtension()
+        }
 
         let duckPlayerOnboardingDecider = DefaultDuckPlayerOnboardingDecider()
         add {
